Update pymef3_file.c
This fixes two bugs: "error when reading by sample and the start sample not in the first segment" as well as "minor bug when reading by timestamps".  The first bug occurs when reading by sample, where beyond the first segment the sample_number is not being calculated correctly.  The second bug is a rounding issue where a partially used block is shifted by up to one index in the output array.  See the "Issues" section of pymef for more details of the problem descriptions.