Correction for sample_for_uutc_c().
Fixed logic in sample_for_uutc_c().  The bug is that if the timestamp passed in was in between segments, or after the end of the last segment, then an extrapolation was being done in all cases - this extrapolation should only be done when inside a block.  Past the end of the block boundary, extrapolating samples doesn't make sense.